change score from float to decimal

+++ sites/models.py
@@ -44,7 +44,8 @@ class Site(models.Model):
 class Score(models.Model):
     site = models.ForeignKey(Site)
     metric = models.ForeignKey(Metric)
-    score = models.FloatField()
+    score = models.DecimalField(max_digits=3,
+                                decimal_places=2)
 
     def __str__(self):
         return "%s %s" % (self.site.name, self.metric.name)
